Bugfix in WordPermutationEnumeration (remove(Object) was being called instead of remove(int). Tests added.

+++ src/name/NicolasCoutureGrenier/Maths/Enumerations/WordPermutationEnumeration.java
@@ -89,7 +89,7 @@ public class WordPermutationEnumeration implements Enumeration<Integer[]> {
       for (Integer j = p.getN() - 1; j >= 0; j--) {
         if (p.get(j)) {
           x[pos.get(j)] = nonzeroindices.get(i);
-          pos.remove(j);
+          pos.remove((int)j);
         }
       }
